fix errors from testing account service setup (#38)
# Description
In the process of testing the latest version of this package in the
chain-setup for
[account-service](https://github.com/AmplicaLabs/account-service), some
bugs were found. This PR does the following:

- More readable error message when failing to fund an account properly
in `UserBuilder.build()`
- In `transactions.ts`, when looking for completed transactions we were
subscribing to all chain events via the system RPC, which could get
events from extrinsics not submitted by us. Instead, process the events
inside the transaction subscription itself.
- `schema_id` & `target_hash` had been hard-coded when provisioning
graph public keys; instead, use the detected values
- Make `EventError` more readable